insights/requirements.txt
google-ai-generativelanguage==0.6.6
google-generativeai==0.7.2
pillow==11.1.0
PyDrive==1.3.1
pypdf==4.3.1
qdrant-client==1.13.2
sqlite-fts4==1.0.3
sqlite-utils==3.38
tiktoken==0.9.0
tqdm==4.66.5
